Separated simulator from circuit
The simulator is now its own class to avoid confusion. Grover's algorithm implemented for a specific example.
