+++ QuantumCircuit.py
@@ -7,6 +7,7 @@ Created on Sat Feb  6 15:35:20 2021
 import QuantumRegister
 import numpy as np
 import SquareMatrix as sm
+import Simulator
 
 class QuantumCircuit:
     def __init__(self, size):
@@ -96,9 +97,8 @@ class QuantumCircuit:
 
         """
         low_lim, high_lim = min(gate_info[1:]), max(gate_info[1:])
-        if gate_info[0] == 'cp':
+        if gate_info[0] == 'cp' or gate_info[0]=='ncp':
             low_lim, high_lim = min(gate_info[1:-1]), max(gate_info[1:-1])
-            print(low_lim, high_lim)
         
         available = True
         for i in range(low_lim, high_lim+1): 
@@ -159,226 +159,34 @@ class QuantumCircuit:
         None.
         """
         self.addBigGate(('ccn', control1, control2, qubit))
-    
-    def cNot(self, gate_info):
+
+
+    def ncp(self, bits, phi):
         """
-        Creates an arbitrary sized controlled not gate between two arbitrary qbits.
+        Adds a phase gate controlled by n other qubits
 
         Parameters
         ----------
-        qbit1 : int
-            Position in the circuit of the control qubit
-        qbit2 : int
-            Position in the circuit of the controlled qubit
+        bits : List if ints
+            The control qubits
+        phi : float
+            Rotation parameter
 
         Returns
         -------
-        SparseMatrix
-            The cnot gate entangling the two qubits given.
-        """
-        qbit1, qbit2 = gate_info
-        
-        if qbit1>qbit2:
-            control_bit = np.abs(qbit2-qbit1)
-            controlled_bit = 0
-        else:
-            control_bit = 0
-            controlled_bit = qbit2 - qbit1
-        
-        elements = [(0,0,1)]
-        dimension = 2**(np.abs(qbit2-qbit1)+1)
-        for i in range(1, dimension):
-            if self.bitactive(i, control_bit):
-                col = self.toggle(i, controlled_bit)
-                elements.append((i, col, 1))
-            else: elements.append((i,i,1))
-        return sm.SparseMatrix(dimension, elements)
-    
-    def ccNot(self, gate_info):
-        """
-        Creates a sparsematrix representing the controlled-controlled-not (ccnot or Tiffoli) 
-        gate for the given qubits. Can be applied to any exisiting qubits.
-
-        Parameters
-        ----------
-        control1 : int
-            first qubit controlling the gate
-        control2 : int
-            second qubit controlling the gate
-        qbit3 : int
-            The qubit to be controlled by the other two
-
-        Returns
-        -------
-        SparseMatrix
-            Matrix representation of the gate
+        None.
 
         """
-        control1, control2, qbit3 = gate_info
-        # Reset the values to range from 0 to maxval
-        minval = min(control1, control2, qbit3)
-        maxval = max(control1, control2, qbit3) - minval
-        control1 = control1 - minval
-        control2 = control2 - minval
-        qbit3 = qbit3 - minval
-        
-        # Create elements and calculate dimensions
-        elements = [(0,0,1)]
-        dimension = 2**(maxval+1)
-        
-        # For each possible bit check whether the control qubits are active
-        for i in range(1, dimension):
-            if self.bitactive(i, control1) and self.bitactive(i, control2):
-                # if control qubits are active, calculate the new value and insert into matrix
-                col = self.toggle(i, qbit3)
-                elements.append((i, col, 1))
-            else: elements.append((i,i,1))
-        return sm.SparseMatrix(dimension, elements)
-    
-    def cZ(self, gate_info):
-        """
-        Creates a controlled z gate given 2 qubits
-
-        Parameters
-        ----------
-        gate_info : tuple(int, int)
-            The two gates to control the z
-
-        Returns
-        -------
-        SparseMatrix
-            Representation of the cz gate
-        """
-        qbit1, qbit2 = gate_info
-        
-        elements = [(0,0,1)]
-        dimension = 2**(np.abs(qbit2-qbit1)+1)
-        for i in range(1, dimension):
-            if self.bitactive(i, qbit1) and self.bitactive(i, qbit2):
-                elements.append((i, i, -1))
-            else: elements.append((i,i,1))
-        return sm.SparseMatrix(dimension, elements)
-    
-    def cP(self, gate_info):
-        """
-        Creates a controlled phase gate for the given qubits
-
-        Parameters
-        ----------
-        gate_info : tuple(int, int, float)
-            The information supplied to the gate. Control qubits as ints, float as the phase.
-
-        Returns
-        -------
-        SparseMatrix
-            Representation of the cp gate
-
-        """
-        qbit1, qbit2, phi = gate_info
-        
-        elements = [(0,0,1)]
-        dimension = 2**(np.abs(qbit2-qbit1)+1)
-        for i in range(1, dimension):
-            if self.bitactive(i, qbit1) and self.bitactive(i, qbit2):
-                elements.append((i, i, np.exp(1j*phi)))
-            else: elements.append((i,i,1))
-        return sm.SparseMatrix(dimension, elements)
+        gate_info = ['ncp']
+        gate_info += bits
+        gate_info.append(phi)
+        self.addBigGate(tuple(gate_info))
     
     def cp(self, qbit1, qbit2, phi):
         self.addBigGate(('cp', qbit1, qbit2, phi))
     
     def swap(self, qbit1, qbit2):
         self.addBigGate(('swap', qbit1, qbit2))
-    
-    def Swap(self, gate_info):
-        qbit1, qbit2 = gate_info
-        qbit1 = qbit1 - min(qbit1, qbit2)
-        qbit2 = qbit2 - min(qbit1, qbit2)
-        elements = []
-        dimension = 2**(np.abs(qbit2-qbit1)+1)
-        for i in range(0, dimension):
-            col = i
-            if (self.bitactive(i, qbit1) and not self.bitactive(i, qbit2)) or (not self.bitactive(i, qbit1) and self.bitactive(i, qbit2)):
-                col = self.toggle(self.toggle(i, qbit1), qbit2)
-            elements.append((i, col, 1))
-        return sm.SparseMatrix(dimension, elements)
-    
-    def addLargeGate(self, gate_info):
-        """
-        Helper function for makeMatrices(). Calls the creators for the larger gates based
-        info provided.
-
-        Parameters
-        ----------
-        gate_info : tuple
-            information concerning the gate
-
-        Returns
-        -------
-        operation : SparseMatrix
-            MAtrix representation of the operation for the gates given.
-
-        """
-        if gate_info[0]=='r':
-            operation = self.Rt(complex(gate_info[1]))
-        elif gate_info[0]=='cn':
-            operation = self.cNot(gate_info[1:])
-        elif gate_info[0]=='ccn':
-            operation = self.ccNot(gate_info[1:])
-        elif gate_info[0]=='swap':
-            operation = self.Swap(gate_info[1:])
-        elif gate_info[0]=='cz':
-            operation = self.cZ(gate_info[1:])
-        elif gate_info[0]=='cp':
-            operation = self.cP(gate_info[1:])
-        
-        return operation
-    
-    def makeMatrices(self):
-        # We should make sparse matrix representations of single gates from the beginning.
-        """
-        Creates the matrices that will be applied to the wavevector
-
-        Returns
-        -------
-        bigmats : numpy array
-            list of np matrices that will be applied to the statevector
-
-        """
-        gates = np.array(self.gates, dtype = object).T
-        #debug
-        print('Gates are:')
-        print(gates)
-        
-        bigmats = []
-        for i, slot in enumerate(gates):
-            bigmat = sm.SparseMatrix(1, [(0,0,1)])
-            for j in slot:
-                if type(j)==tuple:
-                    bigmat = self.addLargeGate(j).tensorProd(bigmat)
-                elif j == 's': continue
-                else: bigmat = sm.toSparse(self.singlegates[j]).tensorProd(bigmat)
-            bigmats.append(bigmat)
-        
-        
-        return np.array(bigmats)
-    
-    def Rt(self, theta):
-        """
-        Creates an r gate with the given phase
-
-        Parameters
-        ----------
-        theta : float
-            The angle in radians which the qubit should be rotated by.
-
-        Returns
-        -------
-        SparseMatrix
-            Matrix representation of the r gate.
-
-        """
-        return sm.toSparse(np.array([[1, 0], [0, np.exp(1j*theta)]], dtype=complex))
         
     def simulate(self):
         """
@@ -390,17 +198,8 @@ class QuantumCircuit:
         Planned: any measurements throughout the experiment
         
         """
-        operations = self.makeMatrices()
-        for operation in operations:
-            self.register.Statevec = operation.Apply(self.register.Statevec)
-            print('Current operation:')
-            print(operation.toDense())
-            
-            print('Current statevector is:')
-            print(self.register)
-            print('With statevector')
-            print(self.register.Statevec)
-        
+        self.register = Simulator.Simulator(self.gates, self.register).simulate()
+
     def show(self):
         print('Register defined as:')
         print(self.register)
@@ -413,8 +212,7 @@ class QuantumCircuit:
         print(self.register)
         print('With statevector')
         print(self.register.Statevec)
-        #self.register.measure()
-        
+        self.register.measure()
     
 if __name__ == '__main__':
     circuit = QuantumCircuit(3)
